Keep appgregated statistics in memory for the past X minutes (#17)
All JSON documents produced by bg_mon are aggregated into an array. Every minute the new array is opened and the old one is closed.

In order to store aggregated data efficiently, brotli compression is used. Tests showed that the compression ratio for the 60 points of data is up to 30 and even higher.

Most of the modern browsers support brotli encoding out of the box, therefore the aggregated data, if requested, is always sent compressed, with `Content-Encoding: br` header.

Firefox has the brotli compression disabled for non-https connections, but it can be enabled in the `about:config`.